+++ packages/independent/url-meta/tests/apply_associations/basic.test.mjs
@@ -3,28 +3,28 @@ import { assert } from "@jsenv/assert";
 import { URL_META } from "@jsenv/url-meta";
 
 // composition
-// {
-//   const actual = URL_META.applyAssociations({
-//     url: "file:///foo.js",
-//     associations: {
-//       node: {
-//         "file:///**/*.js": {
-//           foo: true,
-//         },
-//         "file:///foo.js": {
-//           bar: true,
-//         },
-//       },
-//     },
-//   });
-//   const expected = {
-//     node: {
-//       foo: true,
-//       bar: true,
-//     },
-//   };
-//   assert({ actual, expected });
-// }
+{
+  const actual = URL_META.applyAssociations({
+    url: "file:///foo.js",
+    associations: {
+      node: {
+        "file:///**/*.js": {
+          foo: true,
+        },
+        "file:///foo.js": {
+          bar: true,
+        },
+      },
+    },
+  });
+  const expected = {
+    node: {
+      foo: true,
+      bar: true,
+    },
+  };
+  assert({ actual, expected });
+}
 
 try {
   URL_META.applyAssociations({